bot/requirements.txt
asyncpg==0.29.0
aiogram==3.4.1
fluent.runtime==0.3.1
